27/05/1.27
Большие числа теперь доступны (до 10 в степени 100)
Иконка товара затемняется

+++ Assets/Scripts/MenuManager.cs
@@ -17,7 +17,7 @@ public class MenuManager : MonoBehaviour
     public float exitTime;
     public float timeSinceExit;
     public TimeSpan currentTime;
-    private float _reward = 0;
+    private double _reward = 0;
 
     public void Start ()
     {
